#21 work on parsing the os2 table to get information for the font descriptor dictionary

+++ src/UglyToad.PdfPig/Fonts/TrueType/Tables/CMapSubTables/HighByteMappingCMapTable.cs
@@ -77,7 +77,7 @@
                 {
                     int characterCode = (i << 8) + (subHeader.FirstCode + j);
 
-                    var p = data.ReadUnsignedShort();
+                    var p = (int)data.ReadUnsignedShort();
 
                     if (p > 0)
                     {
